+++ src/ebic/crud/groups.py
@@ -4,13 +4,14 @@ from sqlalchemy import func as f
 from sqlalchemy import or_
 
 from ..auth import User
-from ..models.response import DataCollectionGroupSummaryOut
+from ..models.response import DataCollectionGroupSummaryOut, DataCollectionSummaryOut
 from ..models.table import (
     BLSession,
     DataCollection,
     DataCollectionGroup,
     ExperimentType,
     Proposal,
+    Tomogram,
 )
 from ..utils.auth import check_session
 from ..utils.database import Paged, db, paginate, unravel
@@ -65,3 +66,46 @@ def get_collection_groups(
             )
 
     return paginate(check_session(query, user), limit, page, slow_count=True)
+
+
+def get_collections(
+    limit: int,
+    page: int,
+    groupId: Optional[int],
+    search: str,
+    user: User,
+    onlyTomograms: bool,
+) -> Paged[DataCollectionSummaryOut]:
+    sub_with_row = check_session(
+        (
+            db.session.query(
+                f.row_number()
+                .over(order_by=DataCollection.dataCollectionId)
+                .label("index"),
+                *unravel(DataCollection),
+                f.count(Tomogram.tomogramId).label("tomograms"),
+            )
+            .select_from(DataCollection)
+            .join(BLSession, BLSession.sessionId == DataCollection.SESSIONID)
+            .join(Tomogram, isouter=(not onlyTomograms))
+            .group_by(DataCollection.dataCollectionId)
+            .order_by(DataCollection.dataCollectionId)
+        ),
+        user,
+    )
+
+    if groupId is not None:
+        sub_with_row = sub_with_row.filter(
+            groupId == DataCollection.dataCollectionGroupId
+        )
+
+    sub_result = sub_with_row.subquery()
+
+    query = db.session.query(*sub_result.c).filter(
+        or_(
+            sub_result.c.comments.contains(search),
+            search == "",
+        ),
+    )
+
+    return paginate(query, limit, page, slow_count=True)