+++ src/ebic/routes/collections.py
@@ -1,13 +1,9 @@
 from fastapi import APIRouter, Depends
 
-from ..auth import Permissions, User
+from ..auth import Permissions
 from ..crud import collections as crud
-from ..models.response import (
-    DataCollectionSummaryOut,
-    FullMovie,
-    ProcessingJobOut,
-    Tomogram,
-)
+from ..crud.generic import get_ice_histogram_generic
+from ..models.response import FullMovie, ProcessingJobOut, Tomogram
 from ..utils.database import Paged
 from ..utils.dependencies import pagination
 
@@ -17,20 +13,6 @@ router = APIRouter(
     tags=["Data Collections"],
     prefix="/dataCollections",
 )
-
-
-@router.get("", response_model=Paged[DataCollectionSummaryOut])
-def get_collections(
-    groupId: int = None,
-    page: dict[str, int] = Depends(pagination),
-    search: str = "",
-    onlyTomograms: bool = False,
-    user=Depends(User),
-):
-    """List collections belonging to a data collection group"""
-    return crud.get_collections(
-        groupId=groupId, search=search, user=user, onlyTomograms=onlyTomograms, **page
-    )
 
 
 @router.get("/{collectionId}/tomogram", response_model=Tomogram)
@@ -56,3 +38,13 @@ def get_motion_correction(
 ):
     """Get motion correction and tilt alignment data"""
     return crud.get_motion_correction(collectionId=collectionId, **page)
+
+
+@router.get(
+    "/{collectionId}/iceThickness",
+)
+def get_ice_histogram(dataBin: float = 10000, collectionId: int = Depends(auth)):
+    """Get relative ice thickness histogram"""
+    return get_ice_histogram_generic(
+        parent_type="dataCollection", parent_id=collectionId, dataBin=dataBin
+    )