+++ src/ebic/routes/groups.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends
 
 from ..auth import User
 from ..crud import groups as crud
-from ..models.response import DataCollectionGroupSummaryOut
+from ..models.response import DataCollectionGroupSummaryOut, DataCollectionSummaryOut
 from ..utils.database import Paged
 from ..utils.dependencies import pagination
 
@@ -23,4 +23,20 @@ def get_collection_groups(
     """List collection groups belonging to a session"""
     return crud.get_collection_groups(
         session=session, proposal=proposal, search=search, user=user, **page
+    )
+
+
+@router.get(
+    "/{groupId}/dataCollections", response_model=Paged[DataCollectionSummaryOut]
+)
+def get_collections(
+    groupId: int = None,
+    page: dict[str, int] = Depends(pagination),
+    search: str = "",
+    onlyTomograms: bool = False,
+    user=Depends(User),
+):
+    """List collections belonging to a data collection group"""
+    return crud.get_collections(
+        groupId=groupId, search=search, user=user, onlyTomograms=onlyTomograms, **page
     )