+++ tests/autoproc/test_2d_classification.py
@@ -6,7 +6,6 @@ from ..conftest import mock_send
 def test_get_user(mock_permissions, client):
     """Get particle picking data for an autoprocessing program"""
     resp = client.get("/autoProc/56986680/classification")
-    print(resp.json())
     assert resp.status_code == 200
     assert resp.json()["total"] == 5
 
@@ -15,7 +14,6 @@ def test_sort_particles(mock_permissions, client):
     """Get particle picking data for an autoprocessing program and sort by particles
     per class"""
     resp = client.get("/autoProc/56986680/classification")
-    print(resp.json())
     assert resp.status_code == 200
     assert resp.json()["items"][0]["particlesPerClass"] == 60000
 
@@ -24,7 +22,6 @@ def test_sort_class(mock_permissions, client):
     """Get particle picking data for an autoprocessing program and sort by
     class distribution"""
     resp = client.get("/autoProc/56986680/classification?sortBy=class")
-    print(resp.json())
     assert resp.status_code == 200
     assert resp.json()["items"][0]["classDistribution"] == 0.3
 
@@ -33,7 +30,6 @@ def test_sort_resolution(mock_permissions, client):
     """Get particle picking data for an autoprocessing program and sort by estimated
     resolution"""
     resp = client.get("/autoProc/56986680/classification?sortBy=resolution")
-    print(resp.json())
     assert resp.status_code == 200
     assert resp.json()["items"][0]["estimatedResolution"] == 18