+++ CHANGELOG.rst
@@ -3,6 +3,20 @@ Changelog
 ==========
 
 ++++++++++
+v0.3.0 (24/01/2023)
+++++++++++
+
+**Added**
+
+- Frequency data for ice thickness in data collections (:code:`/dataCollections/{id}/iceThickness`) and autoprocessing programs (:code:`/dataCollections/{id}/iceThickness`)
+
+
+**Changed**
+
+- Moves data collection listing from :code:`/dataCollections` to :code:`/dataGroups/{groupId}/dataCollections`
+- Data collection also displays column with index relative to parent data collection group
+
+++++++++++
 v0.2.0 (24/01/2023)
 ++++++++++